diagnostics: Change executor definition for `GetAllFanSpeed`
Change the executor definition for `GetAllFanSpeed` since the returned
value from libec is of type uint16. This will be more aligned with the
implementation of fan diagnostics.

BUG=b:241863950
TEST=(cr) FEATURES=test USE="-cros-debug" emerge-volteer

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4848675
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Ethan Cheng <yycheng@google.com>

+++ diagnostics/cros_healthd/delegate/delegate_impl.cc
@@ -578,7 +578,7 @@ void DelegateImpl::RunFloatingPoint(base::TimeDelta exec_duration,
 
 void DelegateImpl::GetAllFanSpeed(GetAllFanSpeedCallback callback) {
   auto cros_fd = base::ScopedFD(open(ec::kCrosEcPath, O_RDWR));
-  std::vector<uint32_t> fan_rpms;
+  std::vector<uint16_t> fan_rpms;
   std::optional<uint8_t> num_fans = GetNumFans(cros_fd.get());
 
   if (!num_fans.has_value()) {
